co-generator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ML/AI
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.3.1
chromadb==0.4.22
faiss-cpu==1.7.4

# Document processing
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
numpy==1.26.3
python-dotenv==1.0.0
requests==2.31.0
